modify HW1 Require, type()=>info()

+++ src/two_dimensional_vector.h
@@ -5,28 +5,27 @@
 
 using namespace std;
 
-class TwoDimensionalVector{
+class TwoDimensionalVector {
 	public:
-	TwoDimensionalVector(int x, int y): _x(x), _y(y) {}
+	TwoDimensionalVector(double x, double y): _x(x), _y(y) {}
 
-	int getX() {
+	double getX() {
 		return _x;
 	}
 
-	int getY() {
+	double getY() {
 		return _y;
 	}
 
 	double length() {
-		// sqrt(pow(_x, 2) + pow(_y, 2));
 		return sqrt(pow(_x, 2) + pow(_y, 2));
 	}
 
-	int dot(TwoDimensionalVector* vector) {
+	double dot(TwoDimensionalVector* vector) {
 		return vector->getX() * _x + vector->getY() * _y;
 	}
 
-	int cross(TwoDimensionalVector* vector) {
+	double cross(TwoDimensionalVector* vector) {
 		return _x * vector->getY() - _y * vector->getX();
 	}
 
@@ -35,8 +34,8 @@ class TwoDimensionalVector{
 	}
 
 	private:
-	int _x;
-	int _y;
+	double _x;
+	double _y;
 };
 
 #endif
